# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API For Alur Kerja Services",
        "title": "Alur Kerja Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "${PROD_HOSTNAME}",
    "basePath": "/",
    "paths": {
...
        "/jenisdata": {
            "get": {
                "description": "Get list of all Jenis Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data"
                ],
                "summary": "List All Jenis Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web.JenisDataResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new Jenis Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data"
                ],
                "summary": "Create Jenis Data",
                "parameters": [
                    {
                        "description": "Jenis Data Create Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.JenisDataCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.JenisDataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/jenisdata/{id}": {
            "get": {
                "description": "Get Jenis Data detail by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data"
                ],
                "summary": "Get Jenis Data by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Jenis Data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.JenisDataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update existing Jenis Data by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data"
                ],
                "summary": "Update Jenis Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Jenis Data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Jenis Data Update Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.JenisDataUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.JenisDataResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/jenisdataopd": {
            "post": {
                "description": "Create new Jenis Data OPD",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data OPD"
                ],
                "summary": "Create Jenis Data OPD",
                "parameters": [
                    {
                        "description": "Jenis Data OPD Create Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.JenisDataOpdCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.JenisDataOpdResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/jenisdataopd/detail/{id}": {
            "get": {
                "description": "Get Jenis Data OPD detail by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data OPD"
                ],
                "summary": "Get Jenis Data OPD by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Jenis Data OPD ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.JenisDataOpdResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/jenisdataopd/list/{kode_opd}": {
            "get": {
                "description": "Get list of all Jenis Data OPD",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data OPD"
                ],
                "summary": "List All Jenis Data OPD",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kode OPD",
                        "name": "kode_opd",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web.JenisDataOpdResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/jenisdataopd/{id}": {
            "put": {
                "description": "Update existing Jenis Data OPD by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jenis Data OPD"
                ],
                "summary": "Update Jenis Data OPD",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Jenis Data OPD ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Jenis Data OPD Update Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.JenisDataOpdUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.JenisDataOpdResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "web.JenisDataCreateRequest": {
            "type": "object",
            "required": [
                "jenis_data"
            ],
            "properties": {
                "jenis_data": {
                    "type": "string"
                }
            }
        },
        "web.JenisDataOpdCreateRequest": {
            "type": "object",
            "required": [
                "jenis_data",
                "kode_opd",
                "nama_opd"
            ],
            "properties": {
                "jenis_data": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                }
            }
        },
        "web.JenisDataOpdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jenis_data": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                }
            }
        },
        "web.JenisDataOpdUpdateRequest": {
            "type": "object",
            "required": [
                "id",
                "jenis_data",
                "kode_opd",
                "nama_opd"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jenis_data": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                }
            }
        },
        "web.JenisDataResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jenis_data": {
                    "type": "string"
                }
            }
        },
        "web.JenisDataUpdateRequest": {
            "type": "object",
            "required": [
                "id",
                "jenis_data"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jenis_data": {
                    "type": "string"
                }
            }
        },
...
        "web.WebResponse": {
            "description": "Web Response",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}